(platform/linux) fix regression from 5ce505d4
The keyboard mapping workaround necessarily rearranged things that broke input
for some cases on the linux platform. This is a band-aid on that, but we
still need a decent refactor of symtable.lua and a generator that
translates loadkeys -b style into symtable.lua compatible keymaps (char
table, char index, short value)

FossilOrigin-Name: 30589dc1317668b00cc5d20c0ffb6c2dd84d2cdca1cfec1c8fada241d6325e85

+++ src/platform/linux/keycode_xlate.h
@@ -519,7 +519,7 @@ klut[KEY_UNKNOWN] = K_UNKNOWN;
  * that are running, but provide support scripts that implement detection etc.
  *
  */
-static uint32_t lookup_character(uint16_t code, uint16_t modifiers)
+static uint16_t lookup_character(uint16_t code, uint16_t modifiers)
 {
 	if (code > sizeof(alut) / sizeof(alut[0]))
 		return 0;
